requirements.txt
Django==3.2.9
djangorestframework>=3.12.4
Pillow>=8.4.0
django-cors-headers>=3.10.0
django-dotenv>=1.4.2
the-social-network>=0.0.7
psycopg2-binary
whitenoise